chore(manager): hook up routing for the permissions view tabs
